% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SeeWhoWins_fct_fitModel.R
\name{fitModel}
\alias{fitModel}
\title{fitModel}
\usage{
fitModel(data, col_faction, col_wins, col_total_games)
}
\arguments{
\item{data}{data.frame input}

\item{col_faction}{character string representing column of faction labels}

\item{col_wins}{character string representing column of wins}

\item{col_total_games}{character string representing column of total games}
}
\value{
brms model object
}
\description{
fitModel
}
