% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SeeWhoWins_fct_plotModel.R
\name{plotModel}
\alias{plotModel}
\title{plotModel}
\usage{
plotModel(data, model, col_faction, col_wins, col_total_games, plot_dim)
}
\arguments{
\item{data}{data.frame input}

\item{model}{brms model object}

\item{col_faction}{character string representing column of faction labels}

\item{col_wins}{character string representing column of wins}

\item{col_total_games}{character string representing column of total games}

\item{plot_dim}{clientData from session}
}
\value{
plotly object
}
\description{
plotModel
}
