% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SeeWhoWins_fct_showResults.R
\name{showResults}
\alias{showResults}
\title{showResults}
\usage{
showResults(data, model, col_faction, col_wins, col_total_games)
}
\arguments{
\item{data}{data.frame input}

\item{model}{brms model object}

\item{col_faction}{character string representing column of values}

\item{col_wins}{character string representing column of wins}

\item{col_total_games}{character string representing column of total games}
}
\value{
list containing: (i) DT data table, (ii) shiny box
}
\description{
showResults
}
